Assignments/08b._Migrate_Database/pyproject.toml
[tool.poetry]
name = "08b-migrate-database"
version = "0.1.0"
description = ""
authors = ["Nicolas-Mousten <nicolas.mousten@gmail.com>"]
readme = "README.md"
packages = [{include = "08b"}]

[tool.poetry.dependencies]
python = "^3.11"
alembic = "^1.13.1"
sqlalchemy = "^2.0.29"
pymysql = "^1.1.0"
python-dotenv = "^1.0.1"
cryptography = "^42.0.5"
sqlalchemy-utils = "^0.41.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
